feat(nav): fix navigation to scanner activity
